Update detect_stem.R
Updated the detect_stem.R function so that Stem attribute is converted to a numeric binary and saved as an official LasAttribute, so that it can be written to point clouds. This should be useful if people want to write outputs from detect_stem.R and then read and run stem_diameter_profile() on the results, adding a layer of redundancy. Also good for checking stem delineations in CloudCompare etc.

Cheers,
RJLH